ADD: `parent` + `{xyz,rpy}` params to franka_robot.xacro
These can be used if you want to simulate the robot on a mobile base. In these
cases the `$(arm_id)_link0` must not be connected to `world` but instead fixed
to a link on the mobile robot instead. You could generate a custom combined
platform + arm URDF like so:

```xml
<robot>
  <xacro:include filename="$(find franka_description)/robots/common/franka_robot.xacro"/>

  <!-- Instantiate a mobile robot here, which defines a "base_link" -->

  <xacro:franka_robot
         robot="fr3"
         parent="base_link"
         xyz="0 0 0.25"
         rpy="0 0 ${pi/2}">
  </xacro:franka_robot>

</robot>
```